% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{brassicaData-package}
\alias{brassicaData}
\alias{brassicaData-package}
\title{Example data for gsrc}
\description{
Brassica napus data to use in the vignette of __gsrc__.
This package provides raw idat files for two samples,
a raw data object with 140 samples,
position information for the Brassica60K chip,
a SNP name dictionary for the Brassica60K chip and
synteny blocks for the Brassica napus.
}
\details{
Use \code{data(raw_napus)} to load the dataset and explore the options of __gsrc__.
The raw idat files together with the sample sheets are in /extdata.
Get the full path with \code{list.files(system.file("extdata"), full.names = TRUE,
pattern = "idat")}
}

