% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{Dictionary between idat IDs and SNP names}
\format{A data frame with 53,370 rows and 2 variables:
\describe{
  \item{idatID}{8 digit Integer as used in idat file}
  \item{name}{Name of SNPs, Some have suffix "_B", which indicates that one SNP name has two idat IDs.}
}}
\usage{
dictionary
}
\description{
A table to translate the idat IDs into SNP names.
}
\keyword{datasets}

