% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_napus}
\alias{raw_napus}
\title{Raw data}
\format{A raw data object, consisting of seven elements
\describe{
  \item{chr}{Chromosome name for each SNP}
  \item{pos}{Position for each SNP}
  \item{snps}{Name of SNPs, Some have suffix "_B", which indicates that one SNP name has two idat IDs.}
  \item{beads}{Empty element. Bead information can be stored here.}
  \item{sd}{Empty element. Standard deviation information can be stored here.}
  \item{samples}{Sample names. Two names per sample, suffixed with _Grn and _Red.}
  \item{raw}{Matrix containing the raw data values for each input file.}
}}
\usage{
raw_napus
}
\description{
Raw data object.
}
\keyword{datasets}

