% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synteny_blocks}
\alias{synteny_blocks}
\title{Synteny blocks}
\format{A data frame with 53370 rows and 10 variables:
\describe{
  \item{name}{Name of SNPs, Some have suffix "_B", which indicates that one SNP name has two idat IDs.}
  ...
}}
\source{
\url{http://www.sciencedirect.com/science/article/pii/S2352340915000062}
}
\usage{
synteny_blocks
}
\description{
We extracted the collinearity information from
\url{http://www.sciencedirect.com/science/article/pii/S2352340915000062}.
We filtered out small fragments and only maintained the most prominent synteny blocks.
These blocks support the visual analysis of the data.
For more details see the vignette on synteny blocks in the gsrc package.
}
\keyword{datasets}

